fix: sa-token升级

+++ breeze-base/breeze-base-sa-token/src/main/java/com/breeze/boot/satoken/oauth2/password/BreezePasswordGrantTypeHandler.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023, gaoweixuan (breeze-cloud@foxmail.com).
+ * Copyright (c) 2025, gaoweixuan (breeze-cloud@foxmail.com).
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,16 +14,17 @@
  * limitations under the License.
  */
 
-package com.breeze.boot.satoken;
+package com.breeze.boot.satoken.oauth2.password;
 
 import cn.dev33.satoken.SaManager;
-import cn.dev33.satoken.oauth2.config.SaOAuth2ServerConfig;
-import cn.dev33.satoken.oauth2.strategy.SaOAuth2Strategy;
+import cn.dev33.satoken.oauth2.exception.SaOAuth2Exception;
+import cn.dev33.satoken.oauth2.granttype.handler.PasswordGrantTypeHandler;
+import cn.dev33.satoken.oauth2.granttype.handler.model.PasswordAuthResult;
 import cn.dev33.satoken.secure.BCrypt;
 import cn.dev33.satoken.stp.StpUtil;
+import com.breeze.boot.core.exception.BreezeBizException;
 import com.breeze.boot.core.model.UserPrincipal;
 import com.breeze.boot.core.utils.AesUtil;
-import com.breeze.boot.core.utils.Result;
 import com.breeze.boot.log.bo.SysLogBO;
 import com.breeze.boot.log.enums.LogType;
 import com.breeze.boot.log.events.PublisherSaveSysLogEvent;
@@ -34,32 +35,22 @@ import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Import;
 import org.springframework.util.Assert;
 import org.springframework.web.context.request.RequestContextHolder;
 import org.springframework.web.context.request.ServletRequestAttributes;
-import org.springframework.web.servlet.ModelAndView;
 
 import java.time.LocalDateTime;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.function.Function;
 
 import static com.breeze.boot.core.constants.CacheConstants.ROLE_PERMISSION;
-import static com.breeze.boot.log.enums.LogEnum.LogType.CODE;
+import static com.breeze.boot.core.enums.ResultCode.VERIFY_UN_PASS;
+import static com.breeze.boot.log.enums.LogEnum.LogType.LOGIN;
 import static com.breeze.boot.log.enums.LogEnum.Result.FAIL;
 import static com.breeze.boot.log.enums.LogEnum.Result.SUCCESS;
 
-/**
- * oauth令牌配置
- *
- * @author gaoweixuan
- * @since 2024/09/05
- */
 @Slf4j
 @RequiredArgsConstructor
-@Import({AesSecretProperties.class})
-public class SaTokenOauthConfigure {
+public class BreezePasswordGrantTypeHandler extends PasswordGrantTypeHandler {
     private final static String BCRYPT = "{bcrypt}";
 
     private final IUserDetailService userDetailService;
@@ -68,57 +59,37 @@ public class SaTokenOauthConfigure {
 
     private final PublisherSaveSysLogEvent publisherSaveSysLogEvent;
 
-    /**
-     * Sa-Token OAuth2 定制化配置
-     *
-     * @param oauth2Server oauth2服务器
-     */
-    @Autowired
-    public void configOAuth2Server(SaOAuth2ServerConfig oauth2Server) {
-        // 未登录的视图
-        SaOAuth2Strategy.instance.notLoginView = () -> new ModelAndView("login.html");
+    private final Function<HttpServletRequest, Boolean> captchaServiceFunction;
 
-        // 登录处理函数
-        SaOAuth2Strategy.instance.doLoginHandle = (name, pwd) -> {
-            ServletRequestAttributes requestAttributes = (ServletRequestAttributes) RequestContextHolder.getRequestAttributes();
-            Assert.notNull(requestAttributes, "requestAttributes is null");
+    @Override
+    public PasswordAuthResult loginByUsernamePassword(String username, String password) {
+        ServletRequestAttributes requestAttributes = (ServletRequestAttributes) RequestContextHolder.getRequestAttributes();
+        Assert.notNull(requestAttributes, "requestAttributes is null");
 
-            String decodePwd = AesUtil.decryptStr(pwd, this.aesSecretProperties.getAesSecret());
-            UserPrincipal userPrincipal = this.userDetailService.loadUserByUsername(name);
-            if (BCrypt.checkpw(decodePwd, userPrincipal.getPassword().replace(BCRYPT, ""))) {
-                SysLogBO sysLogBO = this.buildLog(requestAttributes.getRequest(), SUCCESS.getCode(), name);
-                this.publisherSaveSysLogEvent.publisherEvent(new SysLogSaveEvent(sysLogBO));
-                SaManager.getSaTokenDao().delete(ROLE_PERMISSION + userPrincipal.getId());
-                StpUtil.login(userPrincipal.getId());
-                return Result.ok();
-            }
-            SysLogBO sysLogBO = this.buildLog(requestAttributes.getRequest(), FAIL.getCode(), name);
+        if (captchaServiceFunction.apply(requestAttributes.getRequest())) {
+            throw new BreezeBizException(VERIFY_UN_PASS);
+        }
+        String decodePwd = AesUtil.decryptStr(password, this.aesSecretProperties.getAesSecret());
+        UserPrincipal userPrincipal = this.userDetailService.loadUserByUsername(username);
+        String pw_hash = BCrypt.hashpw(password, BCrypt.gensalt());
+        if (BCrypt.checkpw(decodePwd, userPrincipal.getPassword().replace(BCRYPT, ""))) {
+            SysLogBO sysLogBO = this.buildLog(requestAttributes.getRequest(), SUCCESS.getCode(), username);
             this.publisherSaveSysLogEvent.publisherEvent(new SysLogSaveEvent(sysLogBO));
-            return Result.fail("账号名或密码错误");
-        };
-
-        // 授权确认视图
-        SaOAuth2Strategy.instance.confirmView = (clientId, scopes) -> {
-            Map<String, Object> map = new HashMap<>();
-            map.put("clientId", clientId);
-            map.put("scope", scopes);
-            return new ModelAndView("confirm.html", map);
-        };
-
-        // 重写 AccessToken 创建策略，返回会话令牌
-        SaOAuth2Strategy.instance.createAccessToken = (clientId, loginId, scopes) -> {
-            log.info("----返回会话令牌");
-            return StpUtil.getOrCreateLoginSession(loginId);
-        };
+            SaManager.getSaTokenDao().delete(ROLE_PERMISSION + userPrincipal.getId());
+            StpUtil.login(userPrincipal.getId());
+            return new PasswordAuthResult(userPrincipal.getId());
+        }
+        SysLogBO sysLogBO = this.buildLog(requestAttributes.getRequest(), FAIL.getCode(), username);
+        this.publisherSaveSysLogEvent.publisherEvent(new SysLogSaveEvent(sysLogBO));
+        throw new SaOAuth2Exception("无效账号密码");
     }
 
     /**
      * 执行日志 业务
      *
      * @param request  请求
-     * @param result 日志结果
+     * @param result   日志结果
      * @param username 参数
-     *
      * @return {@link SysLogBO }
      */
     @SneakyThrows
@@ -128,7 +99,7 @@ public class SaTokenOauthConfigure {
                 .system(userAgent)
                 .logTitle(LogType.USERNAME_LOGIN.getName())
                 .doType(LogType.USERNAME_LOGIN.getCode())
-                .logType(CODE.getCode())
+                .logType(LOGIN.getCode())
                 .result(result)
                 .ip(request.getRemoteAddr())
                 .requestType(request.getMethod())
@@ -138,5 +109,4 @@ public class SaTokenOauthConfigure {
                 .createName(username)
                 .build();
     }
-
 }
